NCI-Agency/anet#2954: Split AssessmentResultsTable.js into several files
